Fixed several corner-case issues in many algorithms.

+++ libft/ft_striteri.c
@@ -12,10 +12,10 @@
 
 #include "libft.h"
 
-void	ft_striteri(char *str, void (*f)(size_t, char *))
+void	ft_striteri(char *str, void (*f)(unsigned int, char *))
 {
-	size_t	str_len;
-	size_t	i;
+	unsigned int	str_len;
+	unsigned int	i;
 
 	str_len = ft_strlen(str);
 	i = 0;
